+++ src/containers/PageConstructor/components/index.ts
@@ -1,0 +1,2 @@
+export type {ConstructorBlocksProps} from './ConstructorBlocks';
+export {ConstructorBlocks} from './ConstructorBlocks';