+++ src/containers/PageConstructor/index.ts
@@ -2,3 +2,4 @@ export type {PageConstructorProps} from './PageConstructor';
 export {PageConstructor} from './PageConstructor';
 export type {PageConstructorProviderProps} from './Provider';
 export {PageConstructorProvider} from './Provider';
+export * from './components';